Am sters niste update uri puse cam random

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/declarations/globals.java
@@ -7,7 +7,7 @@ public class globals {
     public static double yRedGoal=140;
     public static double xBlueGoal=10;
     public static double yBlueGoal=140;
-    public static double alliance=1; //1 e rosu, 0 e albastru
+    public static int alliance=1; //1 e rosu, 0 e albastru
     // stru
 
 }
